refactor: response 타입 분리
