+++ types/response.d.ts
@@ -1,0 +1,9 @@
+export type Response =
+  | {
+      status: "success";
+      data: MovieResponse;
+    }
+  | {
+      status: "fail";
+      data: [];
+    };
